refactor: move type definitions
Going to introduce board and board-column resources so moving the
respective type aliases to their proper locations.
