+++ src/board-columns/interfaces/board-columns.interface.ts
@@ -1,0 +1,7 @@
+// board-columns.interface.ts
+
+import { UUIDString } from '../../common/types';
+
+export type BoardColumnId = UUIDString;
+
+// __EOF__
