+++ src/boards/interfaces/board.interface.ts
@@ -1,7 +1,7 @@
-// task.interface.ts
+// board.interface.ts
 
 import { UUIDString } from '../../common/types';
 
-export type TaskId = UUIDString;
+export type BoardId = UUIDString;
 
 // __EOF__
